feat: 修正信号掩码 ABI 编码
- 使用 SignalNum 统一维护 Linux signal 编号
- 使用 SignalBit 区分 signal 编码位图，按 1 << (signum - 1) 构造
- 将内核和用户态 sigset_t 路径统一为 64 位处理
- 兼容 4 字节 sigsetsize 输入，并裁剪未支持的信号位
- 修正 sigprocmask、sigsuspend、sigtimedwait、sigreturn 的 mask 读写逻辑
- 更新 signal TODO 文档，记录后续 signal 机制待完善项

+++ os/src/syscall/mod.rs
@@ -497,10 +497,10 @@ pub fn syscall(syscall_id: usize, args: [usize; 6]) -> isize {
             args[2] as *mut UserSigAction,
             args[3], // sigsetsize
         ),
-        SYSCALL_SIGPROCMASK => sys_sigprocmask(args[0] as i32, args[1] as *const u32, args[2] as *mut u32, args[3]),
-        SYSCALL_SIGSUSPEND => sys_sigsuspend(args[0] as *const u32, args[1]),
+        SYSCALL_SIGPROCMASK => sys_sigprocmask(args[0] as i32, args[1] as *const u64, args[2] as *mut u64, args[3]),
+        SYSCALL_SIGSUSPEND => sys_sigsuspend(args[0] as *const u64, args[1]),
         SYSCALL_RT_SIGTIMEDWAIT_TIME32 => sys_rt_sigtimedwait_time32(
-            args[0] as *const u32,
+            args[0] as *const u64,
             args[1] as *mut crate::task::SigInfo,
             args[2] as *const OldTimespec32,
             args[3],
